requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
yfinance>=0.2.28
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
ta-lib>=0.4.28
ta>=0.11.0
gymnasium>=0.29.0
tensorboard>=2.14.0
tqdm>=4.65.0
pyyaml>=6.0

